# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Brokers Web API",
        "description": "Esta es la documentación para la API que manejará la información de las propiedades o inmuebles  You can find\n out more about Swagger at\n[http://swagger.io](http://swagger.io) or on\n[irc.freenode.net, #swagger](http://swagger.io/irc/).",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "yocumo1998@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/panel/addresses": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Obtener todas las direcciones",
                "operationId": "addresses",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/coownerships": {
            "get": {
                "tags": [
                    "Coownerships"
                ],
                "summary": "Obtener todas las copropiedades",
                "operationId": "coownerships",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/coownerships/{id}": {
            "get": {
                "tags": [
                    "Coownerships"
                ],
                "summary": "Obtener una copropiedad",
                "operationId": "coownerships/{id}",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de la copropiedad",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/immovables": {
            "get": {
                "tags": [
                    "Immovables"
                ],
                "summary": "Obtener todos los inmuebles",
                "operationId": "immovables",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/immovable-types": {
            "get": {
                "tags": [
                    "ImmovableTypes"
                ],
                "summary": "Obtener todos los tipos de inmuebles",
                "operationId": "immovable-types",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/immovable-types/{id}": {
            "get": {
                "tags": [
                    "ImmovableTypes"
                ],
                "summary": "Obtener un tipo de inmueble",
                "operationId": "immovable-types/{id}",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del tipo de inmueble",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/applicants": {
            "get": {
                "tags": [
                    "Applicant"
                ],
                "summary": "Obtener todas los aplicantes o solicitantes",
                "operationId": "applicants",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/applicants/{id}": {
            "get": {
                "tags": [
                    "Applicant"
                ],
                "summary": "Obtener un aplicante o solicitante",
                "operationId": "applicant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del aplicante o solicitante",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Ingrese el token en formato (Bearer <token>)",
                "name": "api_key",
                "in": "header"
            }
        }
    },
    "security": {
        "bearer": {
            "type": "apiKey",
            "description": "Ingrese el token en formato (Bearer <token>)",
            "name": "api_key",
            "in": "header"
        }
    }
}